Improve grafana and prometheus liveness and readiness probes
Performing liveness/readiness probes on endpoints other than
health/ready can cause unexpected results, and depending on
the Kubernetes version further problems depending on the size of
the response.

In particular, starting from Kubernetes 1.16 the limit of these
responses have been restricted to 10KB [1]. Thus, on Kubernetes 1.16
and beyond, Grafana was never marked as ready with the default maesh
deployment.

[1] https://github.com/kubernetes/kubernetes/pull/76518