Implement TCPPortMapping
Introducing this type allow us to export from the controller the notion of
TCP state table. This implementation relies on Kubernetes ConfigMap to
persist its state.
This implementation enforce allocating ports for a TCP service in a given
range (minPort, maxPort). Which fixes a bug present in the previous
implementation.
Later on, when using this type, we should accept a TCPPortMapper which can:
- Find: Searches for a port which is associated to a given ServiceWithPort.
- Get: Return the ServiceWithPort associated with the given port.
- Add: Add the ServiceWithPort to the first port available.