Cleanup TCP mapping in ShadowServiceManager
When a service gets deleted or updated we want to make sure
we don't reserve a mapping, which limited in size, for ports
that are no longer needed. This commit relies on the
TCPPortMapper.Remove to cleanup port mapping when an update
or a delete gets detected on a user service.