Add TopologyBuilder
The TopologyBuilder builds a Topology object based on the current state of the kubernetes cluster.
It analyzes the Services, Pods, TrafficSplits and TrafficTargets, to determine who can talk
to who and generate the associated graph.
The TopologyBuilder is used by the Provider to generate the dynamic configuration. As the provider
used to more and more complex, introducing this object allow us to split the problem and let
the Provider do a single thing: generate dynamic config from a Topology.