Replace old Providers by the new one
A new command line option has been added to enable or not the ACL mode. In order
to stay compatible with previous versions, if SMI is enabled by ACL flag is not set,
ACL will be enabled so it behave like it used to.

Every entrypoint generated now has the forwardedHeaders.insecure option to true,
so we get X-Forwarded-For headers for TrafficSplit.