refactor: Key based graph and cleanup tests
The Topology structure has been changed to make it seralizable.
Since there was cycling references in the graph, serialization could
not be a direct operation. Instead of having pointers to the parent and
child resources, it now uses Keys (name + namespace).

This change has allowed a rewrite of the unit tests which now uses
fixtures.