perf: improve performances of the CoreDNSSuite
This test suite has been refactored so we no longer just test that we didn't
received and error when patching the corefile but instead also check that the patched corefile
can be used to resolve a .maesh url.

Some code has also been changed in the dns package to improve the stability of the coredns patching.
Starting from now, even if the corefile/custom corefile has already been patched we will re-patch it.
Since version 1.7.0 of CoreDNS, the `upstream` keyword has been removed, so, if a config was already patched for a previous version or CoreDNS but then the user upgraded it, the corefile would have not been repatched. This commit fix it. This fix also helped us speeding up the CoreDNS tests has we no longer need to re-install a complete CoreDNS before each test.

The coredns/coredns:1.2.6 image has been removed as we are not testing anymore that we don't support it.